restructured and fixed pandas
