+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 # UPDATE THIS BEFORE PUBLISHING
-name = "autora-experimentalist-sampler-novelty"
+name = "autora-experimentalist-novelty"
 description = "AutoRA Novelty Experimentalist"
 authors = [{ name = "Sebastian Musslick", email = "sebastian@musslick.de" }]
 dynamic = ["version"]
@@ -21,7 +21,7 @@ dev = [
 
 [project.urls]
 homepage = "http://www.empiricalresearch.ai"
-repository = "https://github.com/AutoResearch/autora-experimentalist-sampler-novelty"
+repository = "https://github.com/AutoResearch/autora-experimentalist-novelty"
 documentation = "https://autoresearch.github.io/autora/"
 
 [build-system]
