+++ tests/test_exp_novelty_sampler.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from autora.experimentalist.sampler.novelty import novelty_sample, novelty_score_sample
+from autora.experimentalist.novelty import novelty_sample, novelty_score_sample
 import numpy as np
 
 # Note: We encourage you to write more functionality tests for your sampler.
@@ -32,7 +32,7 @@ def test_novelty_sample_1D():
     matrix2 = np.array([1, 2, 3])
 
     # reorder matrix1 according to its distances to matrix2
-    reordered_matrix1 = novelty_sample(condition_pool = matrix1,
+    reordered_matrix1 = novelty_sample(conditions = matrix1,
                                         reference_conditions = matrix2,
                                         num_samples = num_samples)
 
@@ -48,7 +48,7 @@ def test_novelty_sample_ND():
     num_samples = 2
 
     # reorder matrix1 according to its distances to matrix2
-    reordered_matrix1 = novelty_sample(condition_pool = matrix1,
+    reordered_matrix1 = novelty_sample(conditions = matrix1,
                                         reference_conditions = matrix2,
                                         num_samples = num_samples)
 
@@ -63,8 +63,10 @@ def test_novelty_score_sample_ND():
     num_samples = 3
 
     # reorder matrix1 according to its distances to matrix2, and obtain distance score
-    (reordered_matrix1, score) = novelty_score_sample(condition_pool=matrix1,
+    new_conditions = novelty_score_sample(conditions=matrix1,
                                         reference_conditions=matrix2,
                                         num_samples=num_samples)
+
+    score = np.array(new_conditions["score"])
 
     assert score[0] > score[1]  and score[1] > score[2]
